backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "morning-classes-check-backend"
version = "0.1.0"
description = "Backend API and data pipeline scaffold for Morning Classes Check"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.7.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.12",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.0",
  "playwright>=1.52.0",
  "pytest>=8.3.0",
  "ruff>=0.8.0",
]

[project.scripts]
mc-release-gate = "app.tools.release_readiness_gate:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
